+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/AudioMediaGroupContent.kt
@@ -1,0 +1,10 @@
+package dev.inmo.tgbotapi.types.message.content.media
+
+import dev.inmo.tgbotapi.types.InputMedia.AudioMediaGroupMemberInputMedia
+import dev.inmo.tgbotapi.types.files.AudioFile
+
+interface AudioMediaGroupContent : MediaGroupContent {
+    override val media: AudioFile
+
+    override fun toMediaGroupMemberInputMedia(): AudioMediaGroupMemberInputMedia
+}