+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/VisualMediaGroupContent.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.types.message.content.media
+
+import dev.inmo.tgbotapi.types.InputMedia.VisualMediaGroupMemberInputMedia
+
+interface VisualMediaGroupContent : MediaGroupContent {
+    override fun toMediaGroupMemberInputMedia(): VisualMediaGroupMemberInputMedia
+}